zabbix agent: do not cache the server's InetAddress internally, allow re-resolve from time to time
In our environment we noticed that microservices using the Zabbix active Agent would lose track
of Zabbix Proxy when we replaced the proxy instance (which makes IP address to change).

The reason is InetAddress caches the resolved IP internally. The appropriate way is let
agent call InetAddress.getByName() often (because internally they cache the results,
and maintain appropriate TTL) - so that environment DNS changes are properly reflected.